Rename building damage to building structural damage  (#583)
* rename

* match length of =
